Extracted common functionality from vector classes
